Use sibling component
